Update AuthController
The signin action generates the Azure AD signin URL, saves the state value generated by the OAuth client, then redirects the browser to the Azure AD signin page.
The callback action is where Azure redirects after the signin is complete. That action makes sure the state value matches the saved value, then users the authorization code sent by Azure to request an access token. It then redirects back to the home page with the access token in the temporary error value. You'll use this to verify that sign-in is working before moving on.